Add basic auth

+++ backend/post/views.py
@@ -15,8 +15,10 @@ import json
 from urllib.parse import urlparse
 import requests
 import uuid
+from node.models import Node
 
 from post.serializers import PostSerializer, PostSerializerGet
+from node.authentication import BasicAuthentication
 
 
 class PostList(ListCreateAPIView):
@@ -24,6 +26,7 @@ class PostList(ListCreateAPIView):
     serializer_class = PostSerializer
     pagination_class = PostPageNumberPagination
     author_id = None
+    basic_auth = BasicAuthentication()
     # permission_classes = (IsAuthorOrReadOnly,)
 
     def get_queryset(self):
@@ -31,7 +34,10 @@ class PostList(ListCreateAPIView):
 
     # get recent posts of author
     def list(self, request, author_id):
-        # check if author exists on local server
+        response = self.basic_auth.remote_request(request)
+        if response:
+            return response
+
         try:
             Author.objects.get(pk=author_id)
 
@@ -41,14 +47,18 @@ class PostList(ListCreateAPIView):
             if page is not None:
                 serializer =  PostSerializerGet(page, many=True, context={'request':request})
                 return self.get_paginated_response(serializer.data)
+
             serializer =  PostSerializerGet(queryset, many=True, context={'request':request})
             return Response(serializer.data, status=200)
         
         except Author.DoesNotExist:
             return Response("Post not found", status=404)
-
     # create a new post
     def create(self, request, author_id):
+        response = self.basic_auth.local_request(request)
+        if response:
+            return response
+
         try:
             author = Author.objects.get(pk=author_id)
         except Author.DoesNotExist:
@@ -74,8 +84,13 @@ class PostList(ListCreateAPIView):
 class PostDetails(APIView):
     # get post
     # permission_classes = [IsAuthenticated]
+    basic_auth = BasicAuthentication()
     
     def get(self, request, post_id, author_id):
+        response = self.basic_auth.remote_request(request)
+        if response:
+            return response
+
         try:
             post = Post.objects.filter(author_id=author_id).get(pk=post_id)
             serializer = PostSerializerGet(post, context={'request':request})
@@ -86,6 +101,10 @@ class PostDetails(APIView):
     #TODO add authentication
     # edit post
     def post(self, request, post_id, author_id):
+        response = self.basic_auth.local_request(request)
+        if response:
+            return response
+
         # if request.data.get('author') != str(author_id):
         #     return Response("You cannot make a post for this URL", status=400)
         try: 
@@ -103,6 +122,10 @@ class PostDetails(APIView):
 
     # delete post
     def delete(self, request, post_id, author_id):
+        response = self.basic_auth.local_request(request)
+        if response:
+            return response
+
         try:
             post = Post.objects.filter(author_id=author_id).get(pk=post_id)
             post.delete()
@@ -113,6 +136,10 @@ class PostDetails(APIView):
 
     # create new post with post_id
     def put(self, request, post_id, author_id):
+        response = self.basic_auth.local_request(request)
+        if response:
+            return response
+
         try: 
             Author.objects.get(pk=author_id)
         except Author.DoesNotExist:
